added some more layouts... SOME
i hope the formating is correct
just added a bunchload of layouts from https://bruhdooh.home.blog/
also note that for cases like capewell i replaced some of the symbols, so that it's only using ',.[]/\; rather than stuff like -= like it's doing in dvorak or capewell. But this shouldn't matter too much since the feel of the alphas matters more (a-z and basic punctuation (',.)). But you can add this as an todo for the future if you so wish.
i also added my custom layout "apinguefyh"- that i've been using for over two years now. i want to switch layouts in the near future and i wanna use this tool without having to relearn qwerty to do so, so i added mine :)